refactor: more identity/intersect utils
