+++ src/shared/intersect.ts
